Took ProcurementUnitItemContent out from ProcurementUnitItem
